+++ shared/index.ts
@@ -26,7 +26,7 @@ import ToggleSwitch from "./Switch/Switch";
 import RadioInput from "./Radio/Radio";
 import ImageUploader from "./ImageUploader/ImageUploader";
 import ConfirmToken from "./ConfirmToken/ConfirmToken";
-import CustomBreadCrumb from "./customBreadCrumb/CustomBreadCrumb";
+import CustomBreadCrumb from "./CustomBreadCrumbs/CustomBreadCrumbs";
 import StarRating from "./starRating/StarRating";
 import BackNavigation from "./backNavigation/BackNavigation";
